requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.13,<4.0
psycopg2-binary>=2.9,<3.0
djangorestframework-simplejwt>=5.2.0,<6.0